+++ examples/timer.rs
@@ -4,7 +4,7 @@ extern crate sdl2;
 use pushrod::render::callbacks::widget_id_for_name;
 use pushrod::render::engine::Engine;
 use pushrod::render::widget::Widget;
-use pushrod::render::widget_config::{CONFIG_COLOR_SECONDARY, CONFIG_PROGRESS};
+use pushrod::render::widget_config::{CONFIG_COLOR_SECONDARY};
 use pushrod::widgets::progress_widget::*;
 use pushrod::widgets::text_widget::TextWidget;
 use pushrod::widgets::timer_widget::*;
@@ -21,44 +21,41 @@ pub fn main() {
         .build()
         .unwrap();
     let mut engine = Engine::new();
-    let mut widget1 = ProgressWidget::new(20, 20, 360, 40);
+    let mut widget1 = ProgressWidget::new(20, 20, 360, 40, 25);
 
     widget1.set_color(CONFIG_COLOR_SECONDARY, Color::RGB(255, 0, 0));
-    widget1.set_numeric(CONFIG_PROGRESS, 25);
 
-    let mut widget2 = ProgressWidget::new(20, 70, 360, 40);
+    let mut widget2 = ProgressWidget::new(20, 70, 360, 40, 50);
 
     widget2.set_color(CONFIG_COLOR_SECONDARY, Color::RGB(255, 0, 0));
-    widget2.set_numeric(CONFIG_PROGRESS, 50);
 
-    let mut widget3 = ProgressWidget::new(20, 120, 360, 40);
+    let mut widget3 = ProgressWidget::new(20, 120, 360, 40, 75);
 
     widget3.set_color(CONFIG_COLOR_SECONDARY, Color::RGB(255, 0, 0));
-    widget3.set_numeric(CONFIG_PROGRESS, 75);
 
     let mut timer = TimerWidget::new(100, true);
     timer.on_timeout(|x, _widgets| {
         let widget1_id = widget_id_for_name(_widgets, String::from("widget1"));
         let widget2_id = widget_id_for_name(_widgets, String::from("widget2"));
         let widget3_id = widget_id_for_name(_widgets, String::from("widget3"));
-        let timer1_pos = (_widgets[widget1_id]
-            .widget
-            .borrow_mut()
-            .get_numeric(CONFIG_PROGRESS)
-            + 1)
-            % 100;
-        let timer2_pos = (_widgets[widget2_id]
-            .widget
-            .borrow_mut()
-            .get_numeric(CONFIG_PROGRESS)
-            + 1)
-            % 100;
-        let timer3_pos = (_widgets[widget3_id]
-            .widget
-            .borrow_mut()
-            .get_numeric(CONFIG_PROGRESS)
-            + 1)
-            % 100;
+//        let timer1_pos = (_widgets[widget1_id]
+//            .widget
+//            .borrow_mut()
+//            .get_numeric(CONFIG_PROGRESS)
+//            + 1)
+//            % 100;
+//        let timer2_pos = (_widgets[widget2_id]
+//            .widget
+//            .borrow_mut()
+//            .get_numeric(CONFIG_PROGRESS)
+//            + 1)
+//            % 100;
+//        let timer3_pos = (_widgets[widget3_id]
+//            .widget
+//            .borrow_mut()
+//            .get_numeric(CONFIG_PROGRESS)
+//            + 1)
+//            % 100;
 
         eprintln!(
             "{}",
@@ -69,18 +66,18 @@ pub fn main() {
                 .is::<ProgressWidget>()
         );
 
-        _widgets[widget1_id]
-            .widget
-            .borrow_mut()
-            .set_numeric(CONFIG_PROGRESS, timer1_pos);
-        _widgets[widget2_id]
-            .widget
-            .borrow_mut()
-            .set_numeric(CONFIG_PROGRESS, timer2_pos);
-        _widgets[widget3_id]
-            .widget
-            .borrow_mut()
-            .set_numeric(CONFIG_PROGRESS, timer3_pos);
+//        _widgets[widget1_id]
+//            .widget
+//            .borrow_mut()
+//            .set_numeric(CONFIG_PROGRESS, timer1_pos);
+//        _widgets[widget2_id]
+//            .widget
+//            .borrow_mut()
+//            .set_numeric(CONFIG_PROGRESS, timer2_pos);
+//        _widgets[widget3_id]
+//            .widget
+//            .borrow_mut()
+//            .set_numeric(CONFIG_PROGRESS, timer3_pos);
     });
 
     engine.setup(500, 180);