+++ src/widgets/progress_widget.rs
@@ -34,6 +34,7 @@ pub struct ProgressWidget {
     system_properties: HashMap<i32, String>,
     callback_registry: CallbackRegistry,
     base_widget: BaseWidget,
+    progress: u8,
 }
 
 /// Creates a new `ProgressWidget`, which draws a progress bar inside a `BaseWidget`.
@@ -42,7 +43,7 @@ impl ProgressWidget {
     /// `Widget` given the `xywh` coordinates, and the `percentage` of fill from 0-100.  The
     /// base color and border colors are set to white and black, respectively.  Use the
     /// `COLOR_SECONDARY` setting to change the color of the fill for the progress bar.
-    pub fn new(x: i32, y: i32, w: u32, h: u32) -> Self {
+    pub fn new(x: i32, y: i32, w: u32, h: u32, progress: u8) -> Self {
         let mut base_widget = BaseWidget::new(x, y, w, h);
 
         base_widget
@@ -60,7 +61,20 @@ impl ProgressWidget {
             system_properties: HashMap::new(),
             callback_registry: CallbackRegistry::new(),
             base_widget,
+            progress,
         }
+    }
+
+    /// Sets the progress for the widget.  Progress value is between 0 and 100.  Anything over
+    /// 100 will just set the progress to 100.
+    pub fn set_progress(&mut self, progress: u8) {
+        if progress > 100 {
+            self.progress = 100;
+        } else {
+            self.progress = progress;
+        }
+
+        self.get_config().set_invalidate(true);
     }
 }
 
@@ -71,24 +85,19 @@ impl Widget for ProgressWidget {
         self.base_widget.draw(c);
 
         let base_color = self.get_color(CONFIG_COLOR_SECONDARY);
-        let progress = (f64::from(self.get_size(CONFIG_SIZE)[0])
-            * (f64::from(self.get_numeric(CONFIG_PROGRESS)) / 100.0)) as u32;
+        let progress_width = (f64::from(self.get_size(CONFIG_SIZE)[0])
+            * (f64::from(self.progress)) / 100.0) as u32;
+
+        eprintln!("{:?}", progress_width);
 
         c.set_draw_color(base_color);
         c.fill_rect(Rect::new(
             self.config.to_x(1),
             self.config.to_y(1),
-            progress,
+            progress_width,
             self.get_size(CONFIG_SIZE)[1] - 2,
         ))
         .unwrap();
-    }
-
-    /// Responds to a screen redraw only if the `CONFIG_PROGRESS` key was changed.
-    fn on_config_changed(&mut self, _k: u8, _v: Config) {
-        if _k == CONFIG_PROGRESS {
-            self.get_config().set_invalidate(true);
-        }
     }
 
     default_widget_functions!();