+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `CONFIG_PADDING` option to `WidgetConfig` class.
 - Added `as_any` to `Widget` classes so that they can be `downcast_ref`-applied. (#235)
 - Added `default_widget_functions` macro to inject default functions for `Widget` structs. (#235)
+- Removed config options from ProgressWidget, now uses `set_progress` to set the progress values.
 
 ## 0.4.13